Replace simple example with a fibonacci computation
In order to get a sense of performance. This shows
that:

1. the Java version is more performant than naive WASM
2. if we take the precompilation of WASM to binary out,
the execution time of the native library is significantly
faster

This means that in a task, there's a real interest in
caching the instance result if a function is going to be
called multiple times.

It also shows that for simple computation, the penalty
of precompiling is higher than the actual execution time.

```
Using memoized fib written in Rust and compiled to WASM
Rust fib(90) = 2880067194370816120
Took 28ms
Using memoized fib written in Java
Java fib(90) = 2880067194370816120
Took 7ms
With external WASM to native compilation
Precompiled Rust fib(90) = 2880067194370816120
Took 0ms

```