Add MD5 hashing sample
This is a rough experiment to perform a more complex operation,
basically hashing a file. This example shows that the task
implementation becomes significantly more complex, because of
the memory interactions.

Both the Java side, requiring manipulation of memory, and the
Rust side, requiring dark magic to get raw pointers, are complex.