Introduce a higher-level memory allocation API
This makes the interaction with the WASM runtime easier.
In Java, a method call which takes a byte array would
be called directly:

 - foo(bytes)

but because WASM uses its own memory access, instead what
we need to do is:

1. allocate a buffer in the WASM memory
2. write the bytes to the buffer
3. invoke `wasm#foo(pointer,len)
4. dispose memory

The new API makes it much simpler by creating an abstraction,
the `MemoryAccess`, which remembers information about the
pointer and the allocated memory length. The memory can be
automatically released after usage.

Similarly, for a method which _returns_ a byte[], it actually
returns a _pointer_ to the byte array in memory.
So this API makes it easier by wrapping the returned value
into a `MemoryAccess`.