Introduce an annotation processor
The goal of this annotation processor is to demonstrate how we
could simplify invocation of tasks with code generation: the
task ceremony is limited to declaring an interface, annotated
with `@WasmProtocol` which basically defines the inputs/outputs
of a task.

The annotation processor will then generate the task implementation,
and the actual code execution is delegated to the WASM runtime.

This is completely experimental at this stage, and not super
useful since the WASM runtime cannot handle files at this
stage.